+++ src/com/genir/aitweaks/core/features/shipai/AutofireManager.kt
@@ -7,9 +7,12 @@ import com.genir.aitweaks.core.features.shipai.autofire.SyncState
 import com.genir.aitweaks.core.utils.Interval
 import com.genir.aitweaks.core.utils.defaultAIInterval
 import com.genir.aitweaks.core.utils.extensions.customAI
+import com.genir.aitweaks.core.utils.extensions.isMissile
 import com.genir.aitweaks.core.utils.extensions.isPD
+import lunalib.lunaSettings.LunaSettings
 
 class AutofireManager(val ship: ShipAPI) {
+    private val enabledStaggeredFire: Boolean = LunaSettings.getBoolean("aitweaks", "aitweaks_enable_staggered_fire") == true
     private val updateInterval: Interval = defaultAIInterval()
     private val weaponSyncMap: MutableMap<String, SyncState> = mutableMapOf()
 
@@ -24,21 +27,35 @@ class AutofireManager(val ship: ShipAPI) {
     }
 
     private fun ensureAutofire() {
-        // Ensure autofire on all groups.
-        (ship as Ship).setNoWeaponSelected()
-        ship.weaponGroupsCopy.forEach { it.toggleOn() }
+        ship.weaponGroupsCopy.forEach { group ->
+            val weapons = group.weaponsCopy
+
+            val shouldAutofire = when {
+                weapons.any { it.isPD } -> true
+                !ship.hullSpec.isBuiltInMod("missile_reload") && weapons.any { it.isMissile } -> false
+                else -> true
+            }
+
+            if (shouldAutofire) {
+                // Deselect group that should be autofiring.
+                if (ship.selectedGroupAPI == group) (ship as Ship).setNoWeaponSelected()
+                group.toggleOn()
+            } else {
+                group.toggleOff()
+            }
+        }
     }
 
     /** Ensure all weapons that are to fire in staggered mode share an up-to date state. */
     private fun updateWeaponSync() {
+        if (!enabledStaggeredFire) return
+
         val weapons: Sequence<WeaponAPI> = ship.weaponGroupsCopy.flatMap { it.weaponsCopy }.asSequence()
         val syncWeapons = weapons.filter {
             when {
                 it.isBeam -> false
-
                 it.isPD -> false
-
-                it.type == WeaponAPI.WeaponType.MISSILE -> false
+                it.isMissile -> false
 
                 else -> true
             }