+++ src/com/genir/aitweaks/core/features/shipai/Preset.kt
@@ -12,7 +12,7 @@ class Preset {
         const val backoffUpperThreshold = 0.75f
         const val forceVentThreshold = backoffUpperThreshold * 0.6f // will force vent
         const val backoffLowerThreshold = 0.1f
-        const val holdFireThreshold = 0.9f
+        const val holdFireThreshold = 0.93f
         const val damageHistoryDuration = 4f
 
         // Idle time calculation